prediction-api/requirements.txt
Flask
gunicorn
ultralytics
torch
torchvision
google-cloud-storage
opencv-python-headless
numpy